meme_api/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
class-registry==2.1.2
colorama==0.4.6
cors==1.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.42.0
django-auth-ldap==2.1.1
django-cors-headers==3.14.0
django-filter==23.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
filelock==3.10.7
filters==1.3.2
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-ldap==3.4.3
python3-openid==3.2.0
pytz==2023.3
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
urllib3==1.26.15
zope.event==4.6
zope.interface==6.0
